requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql-connector-python==8.4.0
pyTelegramBotAPI==4.16.1
redis==5.0.2
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
telebot==0.0.5
tqdm==4.66.4
urllib3==2.2.1
